// client/src/model/Movie.ts
export class Movie {
    id: number;
    adult: Boolean;
    backdrop_path: string;
    genre_ids: Array<any>;
    original_language: string;
    original_title: string;
    overview: string;
    popularity: number;
    poster_path: string;
    release_date: string;
    title: string;
    video: Boolean;
    vote_average: number;
    vote_count: number;
    isFavorite: Boolean;
  }
  
  